+++ packages/react-native/ReactAndroid/src/main/jni/first-party/yogajni/jni/common.cpp
@@ -18,8 +18,7 @@ void registerNatives(
 
   assertNoPendingJniExceptionIf(env, !clazz);
 
-  auto result =
-      env->RegisterNatives(clazz, methods, static_cast<int32_t>(numMethods));
+  auto result = env->RegisterNatives(clazz, methods, numMethods);
 
   assertNoPendingJniExceptionIf(env, result != JNI_OK);
 }
